crawler/requirements.txt
# OliveYoung Crawler Dependencies
playwright==1.40.0
supabase==2.3.0
python-dotenv==1.0.0
